requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-weaviate
weaviate-client
sentence-transformers
python-dotenv
pypdf
docx2txt
